+++ Student/src/main/java/com/example/naman/repository/StudentRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.naman.entity.Student;
 
 
-public interface StudentRepository extends JpaRepository<Student, Long>
+public interface StudentRepository extends JpaRepository<Student, Integer>
 {
 	List<Student> findByCourseStartsWith(String course);
 }
